compnetized all the chatsw elemnts
