+++ src/components/chat/ChatData.ts
@@ -1,0 +1,156 @@
+import { ChatItem } from './ChatSidebar';
+import { Message } from './ChatMessages';
+
+// Sample chat list data
+export const dummyChats: ChatItem[] = [
+  {
+    id: "1",
+    name: "Alex Johnson",
+    avatar: "https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?ixlib=rb-1.2.1&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80",
+    lastMessage: "Hey, how's your stay going?",
+    time: "5m",
+    unread: true,
+    online: true,
+  },
+  {
+    id: "2",
+    name: "Emma Wilson",
+    avatar: "https://images.unsplash.com/photo-1438761681033-6461ffad8d80?ixlib=rb-1.2.1&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80",
+    lastMessage: "The local festival is tomorrow!",
+    time: "30m",
+    unread: true,
+    online: false,
+  },
+  {
+    id: "3",
+    name: "Michael Chen",
+    avatar: "https://images.unsplash.com/photo-1500648767791-00dcc994a43e?ixlib=rb-1.2.1&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80",
+    lastMessage: "I'll send you directions to the cafe.",
+    time: "2h",
+    unread: true,
+    online: true,
+  },
+  {
+    id: "4",
+    name: "Sarah Lopez",
+    avatar: "https://images.unsplash.com/photo-1494790108377-be9c29b29330?ixlib=rb-1.2.1&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80",
+    lastMessage: "Did you enjoy the local food?",
+    time: "5h",
+    unread: false,
+    online: false,
+  },
+  {
+    id: "5",
+    name: "David Kim",
+    avatar: "https://images.unsplash.com/photo-1507003211169-0a1dd7228f2d?ixlib=rb-1.2.1&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80",
+    lastMessage: "Checking in about your reservation.",
+    time: "1d",
+    unread: false,
+    online: false,
+  },
+  {
+    id: "6",
+    name: "Lisa Wang",
+    avatar: "https://images.unsplash.com/photo-1544005313-94ddf0286df2?ixlib=rb-1.2.1&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80",
+    lastMessage: "I loved the cooking class recommendation!",
+    time: "2d",
+    unread: false,
+    online: false,
+  },
+  {
+    id: "7",
+    name: "John Davis",
+    avatar: "https://images.unsplash.com/photo-1500648767791-00dcc994a43e?ixlib=rb-1.2.1&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80",
+    lastMessage: "When does the trek start tomorrow?",
+    time: "3d",
+    unread: false,
+    online: true,
+  },
+];
+
+// Sample messages data
+export const dummyMessages: Message[] = [
+  {
+    id: "1",
+    sender: "Alex Johnson",
+    senderAvatar: "https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?ixlib=rb-1.2.1&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80",
+    content: "Hey, how's your stay going?",
+    timestamp: "10:30 AM",
+    isSelf: false,
+  },
+  {
+    id: "2",
+    sender: "You",
+    content: "It's great! I'm enjoying the local cuisine a lot.",
+    timestamp: "10:32 AM",
+    isSelf: true,
+  },
+  {
+    id: "3",
+    sender: "Alex Johnson",
+    senderAvatar: "https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?ixlib=rb-1.2.1&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80",
+    content: "That's awesome! Have you tried the local specialty dish yet?",
+    timestamp: "10:33 AM",
+    isSelf: false,
+  },
+  {
+    id: "4",
+    sender: "You",
+    content: "Not yet, but it's on my list. Any specific recommendations?",
+    timestamp: "10:35 AM",
+    isSelf: true,
+  },
+  {
+    id: "5",
+    sender: "Alex Johnson",
+    senderAvatar: "https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?ixlib=rb-1.2.1&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80",
+    content: "You should definitely try 'momos' from the small shop near the market. They're amazing! Also, there's a local festival happening tomorrow if you're interested.",
+    timestamp: "10:40 AM",
+    isSelf: false,
+  },
+  {
+    id: "6",
+    sender: "You",
+    content: "That sounds perfect! What time does the festival start?",
+    timestamp: "10:42 AM",
+    isSelf: true,
+  },
+  {
+    id: "7",
+    sender: "Alex Johnson",
+    senderAvatar: "https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?ixlib=rb-1.2.1&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80",
+    content: "It starts around 4 PM and goes until late night. There will be traditional dances and food stalls. I can send you the exact location if you want.",
+    timestamp: "10:45 AM",
+    isSelf: false,
+  },
+  {
+    id: "8", 
+    sender: "You",
+    content: "That sounds perfect! I would love to attend. Could you send me the location?",
+    timestamp: "10:48 AM",
+    isSelf: true,
+  },
+  {
+    id: "9",
+    sender: "Alex Johnson",
+    senderAvatar: "https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?ixlib=rb-1.2.1&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80",
+    content: "Sure! Here's the location: Town Square near the central market. It's about a 15-minute walk from your homestay. You can't miss it - there will be decorations and music!",
+    timestamp: "10:50 AM",
+    isSelf: false,
+  },
+  {
+    id: "10",
+    sender: "You",
+    content: "Thanks for the information! I'm looking forward to experiencing the local culture.",
+    timestamp: "10:52 AM",
+    isSelf: true,
+  },
+  {
+    id: "11",
+    sender: "Alex Johnson",
+    senderAvatar: "https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?ixlib=rb-1.2.1&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80",
+    content: "You're welcome! Let me know if you need any other recommendations during your stay. I'm happy to help!",
+    timestamp: "10:55 AM",
+    isSelf: false,
+  }
+];
