+++ src/components/chat/index.ts
@@ -1,0 +1,7 @@
+export { default as ChatContainer } from './ChatContainer';
+export { default as ChatSidebar } from './ChatSidebar';
+export { default as ChatHeader } from './ChatHeader';
+export { default as ChatMessages } from './ChatMessages';
+export { default as MessageInput } from './MessageInput';
+export { default as EmptyChat } from './EmptyChat';
+export { dummyChats, dummyMessages } from './ChatData';
